refactor: extract basic-flows flows
